Add schema file.
