+++ pkg/refactor/core/core.go
@@ -1,0 +1,1 @@
+package core
