+++ pkg/refactor/datastore/file_annotations_test.go
@@ -2,6 +2,7 @@ package datastore_test
 
 import (
 	"context"
+	"github.com/direktiv/direktiv/pkg/refactor/database"
 	"reflect"
 	"strings"
 	"testing"
@@ -10,12 +11,11 @@ import (
 	"github.com/direktiv/direktiv/pkg/refactor/datastore/datastoresql"
 	"github.com/direktiv/direktiv/pkg/refactor/filestore"
 	"github.com/direktiv/direktiv/pkg/refactor/filestore/filestoresql"
-	"github.com/direktiv/direktiv/pkg/refactor/utils"
 	"github.com/google/uuid"
 )
 
 func Test_sqlFileAnnotationsStore_SetAndGet(t *testing.T) {
-	db, err := utils.NewMockGorm()
+	db, err := database.NewMockGorm()
 	if err != nil {
 		t.Fatalf("unepxected NewMockGorm() error = %v", err)
 	}
