+++ pkg/refactor/datastore/logstore_test.go
@@ -3,18 +3,18 @@ package datastore_test
 import (
 	"context"
 	"fmt"
+	"github.com/direktiv/direktiv/pkg/refactor/database"
 	"math/rand"
 	"testing"
 	"time"
 
 	"github.com/direktiv/direktiv/pkg/refactor/datastore/datastoresql"
 	"github.com/direktiv/direktiv/pkg/refactor/logengine"
-	"github.com/direktiv/direktiv/pkg/refactor/utils"
 	"github.com/google/uuid"
 )
 
 func Test_Add_Get(t *testing.T) {
-	db, err := utils.NewMockGorm()
+	db, err := database.NewMockGorm()
 	if err != nil {
 		t.Fatalf("unepxected NewMockGorm() error = %v", err)
 	}
